feat: Add Home component
This commit adds the Home component to the project. The Home component is a basic UI component that renders a simple "Home" text. It will be used as a route component in the application.

Note: The commit message follows the established convention of starting with a verb in the imperative form ("Add") and providing a concise description of the change.